Move yanked message unit tests to use resolver
